implement of add and edit product

+++ components/UserProductList.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { View, Text, StyleSheet, FlatList } from "react-native";
+import { View, Text, StyleSheet, FlatList, Alert } from "react-native";
 import ProductItem from "./ProductItem";
 import { useDispatch } from "react-redux";
 import MainButton from "./MainButton";
@@ -8,23 +8,36 @@ import * as productActions from "../store/action/product";
 const UserProductList = (props) => {
     const dispatch = useDispatch();
     const deleteHandler = (id) => {
-        dispatch(productActions.deleteProduct(id));
+        Alert.alert("Are you sure?", "Do you really want to delete this item", [
+            { text: "No", style: "default" },
+            {
+                text: "Yes",
+                style: "destructive",
+                onPress: () => {
+                    dispatch(productActions.deleteProduct(id));
+                },
+            },
+        ]);
     };
 
-    const selectItemHander = (id, title) => {
+    const selectItemHander = (id) => {
         props.navigation.navigate({
-            routeName: "productDetail",
+            routeName: "userEdit",
             params: {
                 productId: id,
-                productTitle: title,
             },
         });
     };
 
     const renderItem = (itemData) => {
         return (
-            <ProductItem onSelectItem={selectItemHander} item={itemData.item}>
-                <MainButton onPress={() => {}}>EDIT</MainButton>
+            <ProductItem
+                onSelectItem={() => selectItemHander(itemData.item.id)}
+                item={itemData.item}
+            >
+                <MainButton onPress={() => selectItemHander(itemData.item.id)}>
+                    EDIT
+                </MainButton>
 
                 <MainButton
                     onPress={() => {
